Added Features and Warnings
- Added a feature were malware metadata values can be used to query the
MITRE ATTC&K database to find the associated software codes. Any that
are not already present as will be added as mitre_att metadata values.
This will not remove existing mitre_att metadata with software values.

- Added a warning check for when info|exploit|technique|tool|malware
key is matching the category key value is not present. While not
required it is a best practice to include the metadata with the
additional information i.e. malware: "malware name"

- Added a warning check for metadata keys that are similarly named to
keys in the standard, this is to try and help catch things like hash1,
hash2, or hash_256. This is a first pass and should be updated with more
common items over time.