Remove global Regex flags from allowed values
The global regex flags, while allowed per-regex, are required to be at
the start of a regex expression. The current code joins regular
extressions resulting in an exception being generated as the global
flags get nested within the expression itself. To avoid this issue, the
global case-insensitive flags can be ommited and instead specify the
case-sensitive equivalence.